refactor: split onboarding presentation primitives
